\name{COGleader}
\alias{COGleader}
\alias{Dimensionality Reduction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Dimension Reduction using Cluster and Lead Lag Analysis
}
\description{
Insert Here a small description}
\usage{
COGleader(Data, NumbClust, type = "STS", TITLE = "", freq = 1/252)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{Object of class yuima.data}
  \item{NumbClust}{Number of Cluster}
  \item{type}{Dissimilarity measure. Default \code{type = "STS"}.}
  \item{TITLE}{...}
  \item{freq}{Default 1/252}
}
\value{Object of class yuima.data}
\author{...}
\examples{
# Insert Here examples
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
